ARCH-3: Refactored RecomendationFilter: replaced custom ItemsList with SingleCheckedRadioButton
