--- WineUp/WineUp/Models/RecomendedFilter+ItemsList+Item.swift
@@ -1,21 +1,0 @@
-//
-//  RecomendedFilter+ItemsList+Item.swift
-//  WineUp
-//
-//  Created by Александр Пахомов on 07.10.2020.
-//
-
-import Foundation
-
-extension RecommendationFilter.ItemsList {
-    /// 
-    struct Item: Identifiable {
-        var id = UUID()
-        ///
-        var sortOrder: CatalogSortOrder
-    }
-}
-
-extension RecommendationFilter.ItemsList.Item: Equatable {
-
-}
