+++ WineUp/WineUp/Views/Filters/RecommendationFilter/RecommendationFilter+ViewModel.swift
@@ -12,19 +12,34 @@ import Foundation
 extension RecommendationFilter {
     final class ViewModel: ObservableObject {
 
+        @Published var catalogSortOrderItems: [CatalogSortOrderItem] = []
+        @Published var checkedCatalogSortOrderItem: CatalogSortOrderItem?
+
+        init() {
+            initStaticCatalogSortOrderItems()
+        }
+    }
+}
+
+extension RecommendationFilter {
+    struct CatalogSortOrderItem: Equatable {
+        var sortOrder: CatalogSortOrder
     }
 }
 
 // MARK: - Public Methods
 
 extension RecommendationFilter.ViewModel {
-    var listViewModel: RecommendationFilter.ItemsList.ViewModel {
-        return .init()
-    }
+
 }
 
 // MARK: - Helpers
 
 private extension RecommendationFilter.ViewModel {
-
+    func initStaticCatalogSortOrderItems() {
+        catalogSortOrderItems = [
+            .init(sortOrder: .recommended),
+            .init(sortOrder: .baseedOnRating)
+        ]
+    }
 }
